Add a stats module to dttools
This is a simple module for keeping counts of things happening in a program. Unless explicitly enabled, nothing is tracked. Note that there is no API for retrieving the stats values, since this is not general-purpose storage. The idea is to instrument interesting parts of the program. In Parrot's read syscall handler, for example, you might put

    stats_inc("parrot.syscall.read", 1);

to increment that counter by one. Counter names are freeform strings, and get initialized automatically. At the end of the program, you can dump out a grep-able table of counts. For events that have some associated value, e.g. the size of a read, you can use

    stats_log("parrot.syscall.readsize", size);

which will record log2(size), giving a histogram with logarithmic bins.